separate registry
